lt01xxx: disable samsung_extdisp driver and regenerate defconfigs
The samsung_extdisp driver for GT-I9500 was affected by CVE-2015-1800 and

CVE-2015-1801. The Galaxy S3 and (lt01xxx) uses an older version of this driver

and is at least affected by CVE-2015-1801. Newer kernel versions for the GT-I9500

had the driver completely removed. I also found no indication that the

driver is actually needed for the S3 (or lt01xxx), so let's disable it.

vulnerability disclosure and further information:

http://blog.quarkslab.com/kernel-vulnerabilities-in-the-samsung-s4.html